Implement Issue #61
